% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBPI.R
\name{getBPI}
\alias{getBPI}
\title{Downloads specimens records for input study BPI accession number}
\usage{
getBPI(BPI, sep = "; ")
}
\arguments{
\item{BPI}{a single study BPI or a vector of study BPIs}

\item{sep}{seperator how to collapse output literature references}
}
\value{
an object of class \code{data.frame} with studies
}
\description{
Downloads specimens records for input study BPI accession number
}
\author{
Franz-Sebastian Krah
}
