% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStudy.R
\name{getStudy_single}
\alias{getStudy_single}
\title{Downloads studies for input study IDs}
\usage{
getStudy_single(id = 42264)
}
\arguments{
\item{id}{a single study id or a vector of study ids}
}
\value{
an object of class \code{list} with studies
}
\description{
Downloads studies for study IDs which are output of associations
}
\author{
Franz-Sebastian Krah
}
