\name{rusda-package}
\alias{rusda-package}
%\alias{rusda}
\docType{package}
\title{
Interface to USDA Databases
}
\description{An interface to the web service methods provided by the United States Department of Agriculture (USDA). The Agricultural Research Service (ARS) provides a large set of databases. The current version of the package holds interfaces to the Systematic Mycology and Microbiology Laboratory (SMML), which consists of four databases: Fungus-Host Distributions, Specimens, Literature and the Nomenclature database. It provides functions for querying these databases. The main function is \code{associations}, which allows searching for fungus-host combinations.}
\details{
\tabular{ll}{
Package: \tab rusda\cr
Type: \tab Package\cr
Version: \tab 1.0.7\cr
Date: \tab 2016-01-20\cr
}

}
\author{
Franz-Sebastian Krah \cr
Maintainer: Franz-Sebastian Krah <f.krah@mailbox.org>
}
\references{
Farr, D.F., & Rossman, A.Y. Fungal Databases, Systematic Mycology and Microbiology Laboratory, ARS, USDA

 \url{http://nt.ars-grin.gov/sbmlweb/fungi/databases.cfm}, 
 \url{http://www.usda.gov/wps/portal/usda/usdahome}

}
\keyword{ package }
