% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/substrate.R
\name{substrate}
\alias{substrate}
\title{Downloads substrate data from SMML Nomenclature DB}
\usage{
substrate(spec, process = TRUE)
}
\arguments{
\item{spec}{a vector of class \code{character} containing fungal or plant species names}

\item{process}{logical, if \code{TRUE} downloading and extraction process is displayed}
}
\value{
an object of mode \code{list} containing substrate for fungus species
}
\description{
Searches and downloads substrate data from SMML Nomenclature database
}
\details{
Don't be disappointed. Not much data there.
But depends on the study group, so give it try.
}
\examples{
\dontrun{
polyporus <- c("Polyporus_rhizophilus", "Polyporus_squamosus",
 "Polyporus_squamulosus", "Polyporus_varius")
subs.poly <- substrate(polyporus, process=TRUE)
subs.poly
}
}
\author{
Franz-Sebastian Krah
}

