\name{rusda-internal}

\alias{clean_step}
\alias{getCOND}
\alias{getHF}
\alias{getMETA}
\alias{getSYNS}


\title{Internal rusda Functions}
\description{Internal \pkg{rusda} functions.}
\note{
These are internal functions in \code{rusda} and are not intended to be called by the user.
}
\keyword{internal}
